requirements.txt
fastapi
uvicorn
pylint
pytest
mypy
httpx
python-decouple
openai
sphinx
sphinx-rtd-theme
